+++ src/main/java/dev/lambdaurora/spruceui/SpruceUI.java
@@ -1,19 +1,23 @@
 package dev.lambdaurora.spruceui;
 
-import dev.architectury.utils.EnvExecutor;
 import dev.lambdaurora.spruceui.hud.HudManager;
-import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fml.loading.FMLLoader;
 
 @Mod(SpruceUI.MODID)
 public class SpruceUI {
-    public static final String MODID = "spruceui";
+    public static final String MODID = "obsidianui";
+    public static final IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
 
     public SpruceUI() {
         ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> new IExtensionPoint.DisplayTest(() -> IExtensionPoint.DisplayTest.IGNORESERVERONLY, (a, b) -> true));
-        EnvExecutor.runInEnv(Dist.CLIENT, () -> this::onInitializeClient);
+        if (FMLLoader.getDist().isClient()) {
+            this.onInitializeClient();
+        }
     }
 
     public void onInitializeClient() {