+++ src/main/java/dev/lambdaurora/spruceui/event/ResolutionChangeEvent.java
@@ -1,0 +1,14 @@
+package dev.lambdaurora.spruceui.event;
+
+import net.minecraft.client.MinecraftClient;
+import net.minecraftforge.eventbus.api.Event;
+
+public class ResolutionChangeEvent extends Event {
+    private final MinecraftClient client;
+    public ResolutionChangeEvent(MinecraftClient client) {
+        this.client = client;
+    }
+    public MinecraftClient getClient() {
+        return client;
+    }
+}